Move files
- .c files in root folder move to "srcs"
- cub3d.h file moves to "includes"
