+++ gnl/get_next_line.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../cub3d.h"
+#include "../includes/cub3d.h"
 
 static char	*ft_get_nlc_addr(char *s)
 {
